requirements.txt
pysha3
pyethash
numpy
